CMake: Add fuzz support.
Enabled the metadata collection feature, scan heuristics, and all-match
mode when fuzzing in the interest of better code coverage.

Also remove deprecated STREAM command.

+++ unit_tests/testcase.py
@@ -151,11 +151,12 @@ class TestCase(unittest.TestCase):
         """
         print("")
 
-        try:
-            shutil.rmtree(cls.path_tmp)
-            cls.log.info("Removed tmp directory: {}".format(cls.path_tmp))
-        except Exception:
-            cls.log.info("No tmp directory to clean up.")
+        if None == os.getenv("KEEPTEMP"):
+            try:
+                shutil.rmtree(cls.path_tmp)
+                cls.log.info("Removed tmp directory: {}".format(cls.path_tmp))
+            except Exception:
+                cls.log.info("No tmp directory to clean up.")
 
     def setUp(self):
         print("")
